[Rust] Move array related code into a separate module
