+++ rust/arrow/src/lib.rs
@@ -27,10 +27,8 @@
 #![allow(bare_trait_objects)]
 
 pub mod array;
-pub mod array_data;
 pub mod bitmap;
 pub mod buffer;
-pub mod builder;
 pub mod compute;
 pub mod csv;
 pub mod datatypes;
